backend/requirements.txt
Django==5.1.7
djangorestframework==3.16.0
python-dotenv==1.1.1
